+++ src/main/java/io/github/astrapi69/io/annotation/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package provides annotations and classes for importing resources
+ */
+package io.github.astrapi69.io.annotation;