+++ src/main/java/io/github/astrapi69/io/file/filter/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides specific classes that implements the interface java.io.FileFilter
+ */
+package io.github.astrapi69.io.file.filter;