+++ src/main/java/io/github/astrapi69/io/file/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides specific enum and extension classes for file operations
+ */
+package io.github.astrapi69.io.file;