+++ src/main/java/io/github/astrapi69/io/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package provides utility classes for reading serializable objects from files and stream
+ * operations
+ */
+package io.github.astrapi69.io;